report

+++ database/migrations/2019_01_31_150936_create_team_details_table.php
@@ -18,7 +18,7 @@ class CreateTeamDetailsTable extends Migration
             $table->string('nip');
             $table->string('submission_id');
             $table->string('position');     //anggota ke
-            $table->bigInteger('individual_score')->nullable();
+            $table->decimal('individual_score', 8, 4)->nullable();
 
             $table->foreign('team_id')->references('id')->on('teams')->onDelete('CASCADE');
             $table->foreign('nip')->references('id')->on('users')->onDelete('CASCADE');
